update ios/android to 5.1.0, plugin version to 1.0.4
